+++ cate/ds/local.py
@@ -689,7 +689,9 @@ class LocalDataStore(DataStore):
         if os.path.isfile(lock_file):
             os.remove(lock_file)
         if remove_files:
-            shutil.rmtree(os.path.join(self._store_dir, data_source.id), ignore_errors=True)
+            data_source_path = os.path.join(self._store_dir, data_source.id)
+            if os.path.isdir(data_source_path):
+                shutil.rmtree(os.path.join(self._store_dir, data_source.id), ignore_errors=True)
         if data_source in self._data_sources:
             self._data_sources.remove(data_source)
 
@@ -816,6 +818,9 @@ class LocalDataStore(DataStore):
         os.makedirs(self._store_dir, exist_ok=True)
         json_files = [f for f in os.listdir(self._store_dir)
                       if os.path.isfile(os.path.join(self._store_dir, f)) and f.endswith('.json')]
+        unfinished_ds = [f for f in os.listdir(self._store_dir)
+                         if os.path.isfile(os.path.join(self._store_dir, f)) and f.endswith('.lock')]
+        json_files = [f for f in json_files if f.replace('.json', '.lock') not in unfinished_ds]
         self._data_sources = []
         for json_file in json_files:
             try: