+++ cate/ops/subset.py
@@ -30,13 +30,13 @@ Components
 """
 
 import xarray as xr
-import numpy as np
-from shapely.geometry import Point, box, LineString
-from shapely.wkt import loads, dumps
 
 from cate.core.op import op, op_input, op_return
 from cate.core.types import PolygonLike, TimeRangeLike, DatasetLike
 from cate.ops.normalize import adjust_spatial_attrs, adjust_temporal_attrs
+
+from cate.util.opimpl import subset_spatial_impl, subset_temporal_impl
+from cate.util.opimpl import subset_temporal_index_impl
 
 
 @op(tags=['geometric', 'spatial', 'subset'], version='1.0')
@@ -54,125 +54,8 @@ def subset_spatial(ds: xr.Dataset,
     not the polygon itself be masked with NaN.
     :return: Subset dataset
     """
-    # Get the bounding box
     region = PolygonLike.convert(region)
-    lon_min, lat_min, lon_max, lat_max = region.bounds
-
-    # Validate the bounding box
-    if (not (-90 <= lat_min <= 90)) or \
-            (not (-90 <= lat_max <= 90)) or \
-            (not (-180 <= lon_min <= 180)) or \
-            (not (-180 <= lon_max <= 180)):
-        raise ValueError('Provided polygon extends outside of geospatial'
-                         ' bounds: latitude [-90;90], longitude [-180;180]')
-
-    simple_polygon = False
-    if region.equals(box(lon_min, lat_min, lon_max, lat_max)):
-        # Don't do the computationally intensive masking if the provided
-        # region is a simple box-polygon, for which there will be nothing to
-        # mask.
-        simple_polygon = True
-
-    crosses_antimeridian = _crosses_antimeridian(region)
-
-    if crosses_antimeridian and not simple_polygon:
-        # Unlikely but plausible
-        raise NotImplementedError('Spatial subset over the International Date'
-                                  ' Line is currently implemented for simple,'
-                                  ' rectangular polygons only.')
-
-    if crosses_antimeridian:
-        # Shapely messes up longitudes if the polygon crosses the antimeridian
-        lon_min, lon_max = lon_max, lon_min
-
-        # Can't perform a simple selection with slice, hence we have to
-        # construct an appropriate longitude indexer for selection
-        lon_left_of_idl = slice(lon_min, 180)
-        lon_right_of_idl = slice(-180, lon_max)
-        lon_index = xr.concat((ds.lon.sel(lon=lon_right_of_idl),
-                               ds.lon.sel(lon=lon_left_of_idl)), dim='lon')
-        indexers = {'lon': lon_index, 'lat': slice(lat_min, lat_max)}
-        retset = ds.sel(**indexers)
-
-        if mask:
-            # Preserve the original longitude dimension, masking elements that
-            # do not belong to the polygon with NaN.
-            return adjust_spatial_attrs(retset.reindex_like(ds.lon))
-        else:
-            # Return the dataset with no NaNs and with a disjoint longitude
-            # dimension
-            return adjust_spatial_attrs(retset)
-
-    if not mask or simple_polygon:
-        # The polygon doesn't cross the IDL, it is a simple box -> Use a simple slice
-        lat_slice = slice(lat_min, lat_max)
-        lon_slice = slice(lon_min, lon_max)
-        indexers = {'lat': lat_slice, 'lon': lon_slice}
-        return adjust_spatial_attrs(ds.sel(**indexers))
-
-    # Create the mask array. The result of this is a lon/lat DataArray where
-    # all values falling in the region or on its boundary are denoted with True
-    # and all the rest with False
-    lonm, latm = np.meshgrid(ds.lon.values, ds.lat.values)
-    mask = np.array([Point(lon, lat).intersects(region) for lon, lat in
-                     zip(lonm.ravel(), latm.ravel())], dtype=bool)
-    mask = xr.DataArray(mask.reshape(lonm.shape),
-                        coords={'lon': ds.lon, 'lat': ds.lat},
-                        dims=['lat', 'lon'])
-
-    # Mask values outside the polygon with NaN, crop the dataset
-    return adjust_spatial_attrs(ds.where(mask, drop=True))
-
-
-def _crosses_antimeridian(region: PolygonLike.TYPE) -> bool:
-    """
-    Determine if the given region crosses the Antimeridian line, by converting
-    the given Polygon from -180;180 to 0;360 and checking if the antimeridian
-    line crosses it.
-
-    This only works with Polygons without holes
-
-    :param region: PolygonLike to test
-    """
-    region = PolygonLike.convert(region)
-
-    # Retrieving the points of the Polygon are a bit troublesome, parsing WKT
-    # is more straightforward and probably faster
-    new_wkt = 'POLYGON (('
-
-    # [10:-2] gets rid of POLYGON (( and ))
-    for point in dumps(region)[10:-2].split(','):
-        point = point.strip()
-        lon, lat = point.split(' ')
-        lon = float(lon)
-        if -180 <= lon < 0:
-            lon += 360
-        new_wkt += '{} {}, '.format(lon, lat)
-    new_wkt = new_wkt[:-2] + '))'
-
-    converted = loads(new_wkt)
-
-    # There's a problem at this point. Any polygon crossed by the zeroth
-    # meridian can in principle convert to an inverted polygon that is crossed
-    # by the antimeridian.
-
-    if not converted.is_valid:
-        # The polygon 'became' invalid upon conversion => probably the original
-        # polygon is what we want
-        return False
-
-    test_line = LineString([(180, -90), (180, 90)])
-    if test_line.crosses(converted):
-        # The converted polygon seems to be valid and crossed by the
-        # antimeridian. At this point there's no 'perfect' way how to tell if
-        # we wanted the converted polygon or the original one.
-
-        # A simple heuristic is to check for size. The smaller one is quite
-        # likely the desired one
-        if converted.area < region.area:
-            return True
-        else:
-            return False
+    return adjust_spatial_attrs(subset_spatial_impl(ds, region, mask))
 
 
 @op(tags=['subset', 'temporal', 'filter'], version='1.0')
@@ -190,16 +73,7 @@ def subset_temporal(ds: DatasetLike.TYPE,
     """
     ds = DatasetLike.convert(ds)
     time_range = TimeRangeLike.convert(time_range)
-    # If it can be selected, go ahead
-    try:
-        time_slice = slice(time_range[0], time_range[1])
-        indexers = {'time': time_slice}
-        return adjust_temporal_attrs(ds.sel(**indexers))
-    except TypeError:
-        raise ValueError('Time subset operation expects a dataset with the'
-                         ' time coordinate of type datetime64[ns], but received'
-                         ' {}. Running the normalize operation on this'
-                         ' dataset may help'.format(ds.time.dtype))
+    return adjust_temporal_attrs(subset_temporal_impl(ds, time_range))
 
 
 @op(tags=['subset', 'temporal', 'filter', 'utility'], version='1.0')
@@ -217,8 +91,4 @@ def subset_temporal_index(ds: DatasetLike.TYPE,
     :return: Subset dataset
     """
     ds = DatasetLike.convert(ds)
-    # we're creating a slice that includes both ends
-    # to have the same functionality as subset_temporal
-    time_slice = slice(time_ind_min, time_ind_max + 1)
-    indexers = {'time': time_slice}
-    return ds.isel(**indexers)
+    return subset_temporal_index_impl(ds, time_ind_min, time_ind_max)