+++ CHANGES.md
@@ -1,14 +1,23 @@
 ## Changes in version 1.0.0.dev1 (unreleased)
 
 ### Improvements and new Features
+* Global temporal attributes are adjusted automatically when opening new datasets
+* Global temporal attributes are adjusted automatically when converting from data frames
+* Normalization and subsetting operation implementation logic is refactored out to util so that it can be re-used throughout Cate
 
 ### Issues Fixed/Resolved
 
+* Fixed displaying broken/incomplete/canceled data sources on local data sources list
+  [#375](https://github.com/CCI-Tools/cate/issues/375)
 * Generated resource names not always unique
   [#391](https://github.com/CCI-Tools/cate/issues/391)
 * Multiple concurrent attempts to load the ODP index now always return the same result
   [#386](https://github.com/CCI-Tools/cate/issues/386)
-  
+* Use global temporal attributes to determine temporal resolution in aggregation operations
+  [#340](https://github.com/CCI-Tools/cate/issues/340)
+* Only allow valid python identifiers as resource names
+  [#436](https://github.com/CCI-Tools/cate/issues/436)
+
 ## Changes in version 0.9.0.dev7
 
 ### Improvements and new Features